削除と更新のアラートを表示

+++ Sources/AppFeature/UI/List/ListViewModel.swift
@@ -5,6 +5,16 @@
 import Foundation
 import UIKit
 
+// MARK: - Action
+
+enum ListViewAction {
+    case load
+    case reload
+    case onDeleteButtonClick(offsets: IndexSet)
+    case onDelete
+    case onDeleteAlertDismiss
+}
+
 @MainActor
 final class ListViewModel: ObservableObject {
     @Published private(set) var uiState = ListUiState()
@@ -13,27 +23,35 @@ final class ListViewModel: ObservableObject {
         self.uiState = uiState
     }
 
-    func loadItems() {
-        Task {
-            uiState.items = try await ItemRepository().load()
+    func send(_ action: ListViewAction) {
+        switch action {
+        case .load:
+            Task {
+                uiState.items = try await ItemRepository().load()
+            }
+        case .reload:
+            send(.load)
+        case let .onDeleteButtonClick(offsets: offsets):
+            print(offsets.map { uiState.items[$0] })
+            uiState.deleteItems = offsets.map { uiState.items[$0] }
+            uiState.deleteAlertMessage = "削除しますか？"
+        case .onDelete:
+            if let deleteItems = uiState.deleteItems {
+                deleteItems.forEach { item in
+                    Task {
+                        try await ItemRepository().delete(item: item)
+                        send(.reload)
+                    }
+                }
+            } else {
+                fatalError("no deleteItems")
+            }
+        case .onDeleteAlertDismiss:
+            uiState.deleteAlertMessage = nil
         }
     }
 
     func loadImage(fileName: String?) -> UIImage? {
         ImageRepository().load(fileName: fileName)
-    }
-
-    func reload() {
-        loadItems()
-    }
-
-    func delete(offsets: IndexSet) {
-        let items = offsets.map { uiState.items[$0] }
-        items.forEach { item in
-            Task {
-                try await ItemRepository().delete(item: item)
-                reload()
-            }
-        }
     }
 }
